+++ CarritoMVC/Controllers/EmpleadosController.cs
@@ -22,31 +22,55 @@ namespace CarritoMVC.Controllers
         // GET: Empleados
         public async Task<IActionResult> Index()
         {
-              return View(await _context.Empleados.ToListAsync());
+            if (Login())
+            {
+                return View(await _context.Empleados.ToListAsync());
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+              
         }
 
         // GET: Empleados/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null || _context.Empleados == null)
+            if (Login())
             {
-                return NotFound();
-            }
+                if (id == null || _context.Empleados == null)
+                {
+                    return NotFound();
+                }
 
-            var empleado = await _context.Empleados
-                .FirstOrDefaultAsync(m => m.EmpleadoId == id);
-            if (empleado == null)
+                var empleado = await _context.Empleados
+                    .FirstOrDefaultAsync(m => m.EmpleadoId == id);
+                if (empleado == null)
+                {
+                    return NotFound();
+                }
+
+                return View(empleado);
+            }
+            else
             {
-                return NotFound();
+                return RedirectToAction("Index", "Home");
             }
-
-            return View(empleado);
+            
         }
 
         // GET: Empleados/Create
         public IActionResult Create()
         {
-            return View();
+            if (Login())
+            {
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+           
         }
 
         // POST: Empleados/Create
@@ -56,29 +80,45 @@ namespace CarritoMVC.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("EmpleadoId,Telefono,Direccion,Nombre,Apellido,Dni,Email,FechaAlta,Password")] Empleado empleado)
         {
-            if (ModelState.IsValid)
+            if (Login())
             {
-                _context.Add(empleado);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(empleado);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                return View(empleado);
             }
-            return View(empleado);
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+           
         }
 
         // GET: Empleados/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null || _context.Empleados == null)
+            if (Login())
             {
-                return NotFound();
-            }
+                if (id == null || _context.Empleados == null)
+                {
+                    return NotFound();
+                }
 
-            var empleado = await _context.Empleados.FindAsync(id);
-            if (empleado == null)
-            {
-                return NotFound();
+                var empleado = await _context.Empleados.FindAsync(id);
+                if (empleado == null)
+                {
+                    return NotFound();
+                }
+                return View(empleado);
             }
-            return View(empleado);
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            
         }
 
         // POST: Empleados/Edit/5
@@ -88,50 +128,66 @@ namespace CarritoMVC.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("EmpleadoId,Telefono,Direccion,Nombre,Apellido,Dni,Email,FechaAlta,Password")] Empleado empleado)
         {
-            if (id != empleado.EmpleadoId)
+            if (Login())
             {
-                return NotFound();
+                if (id != empleado.EmpleadoId)
+                {
+                    return NotFound();
+                }
+
+                if (ModelState.IsValid)
+                {
+                    try
+                    {
+                        _context.Update(empleado);
+                        await _context.SaveChangesAsync();
+                    }
+                    catch (DbUpdateConcurrencyException)
+                    {
+                        if (!EmpleadoExists(empleado.EmpleadoId))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
+                    }
+                    return RedirectToAction(nameof(Index));
+                }
+                return View(empleado);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
             }
 
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(empleado);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!EmpleadoExists(empleado.EmpleadoId))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            return View(empleado);
         }
 
         // GET: Empleados/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null || _context.Empleados == null)
+            if (Login())
             {
-                return NotFound();
-            }
+                if (id == null || _context.Empleados == null)
+                {
+                    return NotFound();
+                }
 
-            var empleado = await _context.Empleados
-                .FirstOrDefaultAsync(m => m.EmpleadoId == id);
-            if (empleado == null)
+                var empleado = await _context.Empleados
+                    .FirstOrDefaultAsync(m => m.EmpleadoId == id);
+                if (empleado == null)
+                {
+                    return NotFound();
+                }
+
+                return View(empleado);
+            }
+            else
             {
-                return NotFound();
+                return RedirectToAction("Index", "Home");
             }
-
-            return View(empleado);
+           
         }
 
         // POST: Empleados/Delete/5
@@ -139,23 +195,45 @@ namespace CarritoMVC.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            if (_context.Empleados == null)
+            if (Login())
             {
-                return Problem("Entity set 'CarritoContext.Empleados'  is null.");
+                if (_context.Empleados == null)
+                {
+                    return Problem("Entity set 'CarritoContext.Empleados'  is null.");
+                }
+                var empleado = await _context.Empleados.FindAsync(id);
+                if (empleado != null)
+                {
+                    _context.Empleados.Remove(empleado);
+                }
+
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
             }
-            var empleado = await _context.Empleados.FindAsync(id);
-            if (empleado != null)
+            else
             {
-                _context.Empleados.Remove(empleado);
+                return RedirectToAction("Index", "Home");
             }
-            
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+
         }
 
         private bool EmpleadoExists(int id)
         {
           return _context.Empleados.Any(e => e.EmpleadoId == id);
+        }
+
+        public bool Login()
+        {
+            bool l;
+            if (HttpContext.Session.GetString("EmpleadoId") != null && HttpContext.Session.GetString("Admin") == true.ToString())
+            {
+                l = true;
+            }
+            else
+            {
+                l = false;
+            }
+            return l;
         }
     }
 }
