+++ CarritoMVC/Controllers/ProductosController.cs
@@ -22,8 +22,16 @@ namespace CarritoMVC.Controllers
         // GET: Productos
         public ActionResult Index()
         {
-            ViewBag.productos = _context.Productos.Include(p => p.Categoria);
-            return View();
+            if (Login())
+            {
+                ViewBag.productos = _context.Productos.Include(p => p.Categoria);
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            
         }
 
         public ActionResult ProductosXCategoria(int id)
@@ -56,8 +64,16 @@ namespace CarritoMVC.Controllers
         // GET: Productos/Create
         public IActionResult Create()
         {
-            ViewData["CategoriaId"] = new SelectList(_context.Categorias, "CategoriaId", "Descripcion");
-            return View();
+            if (Login())
+            {
+                ViewData["CategoriaId"] = new SelectList(_context.Categorias, "CategoriaId", "Descripcion");
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            
         }
 
         // POST: Productos/Create
@@ -67,31 +83,48 @@ namespace CarritoMVC.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("CategoriaId,Imagen,Nombre,Descripcion,PrecioVigente,Activo,Destacado")] Producto producto)
         {
-            if (ModelState.IsValid)
+            if (Login())
             {
-                _context.Add(producto);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(producto);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                ViewData["CategoriaId"] = new SelectList(_context.Categorias, "CategoriaId", "Descripcion", producto.CategoriaId);
+                return View(producto);
             }
-            ViewData["CategoriaId"] = new SelectList(_context.Categorias, "CategoriaId", "Descripcion", producto.CategoriaId);
-            return View(producto);
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+           
         }
 
         // GET: Productos/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null || _context.Productos == null)
+            if (Login())
             {
-                return NotFound();
+                if (id == null || _context.Productos == null)
+                {
+                    return NotFound();
+                }
+
+                var producto = await _context.Productos.FindAsync(id);
+                if (producto == null)
+                {
+                    return NotFound();
+                }
+                ViewData["CategoriaId"] = new SelectList(_context.Categorias, "CategoriaId", "Descripcion", producto.CategoriaId);
+                return View(producto);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
             }
 
-            var producto = await _context.Productos.FindAsync(id);
-            if (producto == null)
-            {
-                return NotFound();
-            }
-            ViewData["CategoriaId"] = new SelectList(_context.Categorias, "CategoriaId", "Descripcion", producto.CategoriaId);
-            return View(producto);
+           
         }
 
         // POST: Productos/Edit/5
@@ -101,52 +134,68 @@ namespace CarritoMVC.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("ProductoId,CategoriaId,Imagen,Nombre,Descripcion,PrecioVigente,Activo,Destacado")] Producto producto)
         {
-            if (id != producto.ProductoId)
+            if (Login())
             {
-                return NotFound();
-            }
+                if (id != producto.ProductoId)
+                {
+                    return NotFound();
+                }
 
-            if (ModelState.IsValid)
-            {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(producto);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!ProductoExists(producto.ProductoId))
+                    try
                     {
-                        return NotFound();
+                        _context.Update(producto);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
-                        throw;
+                        if (!ProductoExists(producto.ProductoId))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+                ViewData["CategoriaId"] = new SelectList(_context.Categorias, "CategoriaId", "Descripcion", producto.CategoriaId);
+                return View(producto);
             }
-            ViewData["CategoriaId"] = new SelectList(_context.Categorias, "CategoriaId", "Descripcion", producto.CategoriaId);
-            return View(producto);
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            
         }
 
         // GET: Productos/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null || _context.Productos == null)
+            if (Login())
             {
-                return NotFound();
-            }
+                if (id == null || _context.Productos == null)
+                {
+                    return NotFound();
+                }
 
-            var producto = await _context.Productos
-                .Include(p => p.Categoria)
-                .FirstOrDefaultAsync(m => m.ProductoId == id);
-            if (producto == null)
+                var producto = await _context.Productos
+                    .Include(p => p.Categoria)
+                    .FirstOrDefaultAsync(m => m.ProductoId == id);
+                if (producto == null)
+                {
+                    return NotFound();
+                }
+
+                return View(producto);
+            }
+            else
             {
-                return NotFound();
+                return RedirectToAction("Index", "Home");
             }
-
-            return View(producto);
+           
         }
 
         // POST: Productos/Delete/5
@@ -154,23 +203,45 @@ namespace CarritoMVC.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            if (_context.Productos == null)
+            if (Login())
             {
-                return Problem("Entity set 'CarritoContext.Productos'  is null.");
+                if (_context.Productos == null)
+                {
+                    return Problem("Entity set 'CarritoContext.Productos'  is null.");
+                }
+                var producto = await _context.Productos.FindAsync(id);
+                if (producto != null)
+                {
+                    _context.Productos.Remove(producto);
+                }
+
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
             }
-            var producto = await _context.Productos.FindAsync(id);
-            if (producto != null)
+            else
             {
-                _context.Productos.Remove(producto);
+                return RedirectToAction("Index", "Home");
             }
-            
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+
         }
 
         private bool ProductoExists(int id)
         {
           return _context.Productos.Any(e => e.ProductoId == id);
+        }
+
+        public bool Login()
+        {
+            bool l;
+            if (HttpContext.Session.GetString("EmpleadoId") != null && HttpContext.Session.GetString("Admin") == true.ToString())
+            {
+                l = true;
+            }
+            else
+            {
+                l = false;
+            }
+            return l;
         }
     }
 }
